requirements.txt
playwright==1.44.0
beautifulsoup4==4.12.3
httpx==0.27.0
anthropic==0.28.0
python-dotenv==1.0.1
feedparser==6.0.11
jinja2==3.1.4
rich==13.7.1
pytest==8.2.2
pytest-asyncio==0.23.7
